refactor(mso_mdoc): modularize verifier code and cleanup audit labels
Split verifier.py into focused modules (trust_store.py, cred_verifier.py,
pres_verifier.py). Further split pres_verifier.py into mdoc_item.py and
mdoc_verify.py for better separation of concerns.

Also strip C-N and M-N audit report item labels from comments and
docstrings throughout the mso_mdoc plugin, as the referenced report
has been removed.

Changes:
- Split verifier.py → trust_store.py, cred_verifier.py, pres_verifier.py
- Split pres_verifier.py → mdoc_item.py, mdoc_verify.py
- Update all imports to use focused modules
- Remove audit report labels (C-N, M-N prefixes) from comments
