+++ oid4vc/mso_mdoc/mdoc/__init__.py
@@ -1,8 +1,8 @@
 """MDoc module."""
 
 from .issuer import isomdl_mdoc_sign, parse_mdoc
+from .mdoc_verify import MdocVerifyResult, mdoc_verify
 from .utils import extract_signing_cert, flatten_trust_anchors, split_pem_chain
-from .verifier import MdocVerifyResult, mdoc_verify
 
 __all__ = [
     "isomdl_mdoc_sign",
