+++ oid4vc/mso_mdoc/storage/certificates.py
@@ -137,7 +137,7 @@ async def get_certificate_for_key(session: ProfileSession, key_id: str) -> Optio
         if not records:
             return None
 
-        # M-2: take the most recently created certificate to get deterministic,
+        # Take the most recently created certificate for deterministic,
         # reproducible results when multiple certs share the same key_id.
         records.sort(
             key=lambda r: json.loads(r.value).get("created_at", ""),
